flask/requirements-prod.txt
# Production requirements - extends base requirements.txt
-r requirements.txt

# Production-specific packages
gunicorn>=20.1.0
gevent>=21.12.0
psycopg2-binary>=2.9.0

# Security & Monitoring
Werkzeug>=2.0.0

# Environment handling for production
python-dotenv>=0.19.0
